refactor(tests): split integration suites by concern
